chore: og 정보를 추가합니다
