+++ web/app/_shared/components/index.ts
@@ -1,0 +1,2 @@
+export * from './EventListener'
+export * from './Header'
